Make HilbertEncoder::sort generic
